requirements.txt
discord.py==1.5.0
pymongo==4.1.1
requests==2.28.1